\name{BloodPressure}
\alias{BloodPressure}
\docType{data}
\title{Blood Pressure and Heart Rate Readings}
\description{
  These data were collected during June, July, and August, 2006 
  by Taoying Bian.
}
\usage{data(BloodPressure)}
\format{
  A data frame with 15 observations on the following 7 variables:


  \describe{
    \item{month}{month of the year. June 2006 = 6.}
    \item{day}{the day of the month.}
    \item{hour}{the 24-clock hour.}
    \item{minute}{the clock minute of the hour.}
    \item{systolic}{systolic blood pressure reading (in mm Hg).}
    \item{diastolic}{diastolic blood pressure reading (in mm Hg).}
    \item{heart.rate}{heart rate reading, in beats per minute.}
  }
}
\details{
  During June, July, and August 2006, Mrs. Taoying Bian collected data concerning the 
  her blood pressure and heart rate.
}

\keyword{datasets}
\source{
  These data were collected by Taoying Bian during June, July, and August 2006.
}


