\name{IPSUR-data}
\docType{data}
\alias{IPSUR-data}
\title{Objects generated by IPSUR}
\description{
This data set gives everything that was generated by IPSUR during the build process. This includes vectors, data frames, lists, fitted models, et cetera. Note that it does not include objects which were deliberately removed from the workspace for whatever reason.
}
\usage{data(IPSUR)}
\format{A workspace saved at the conclusion of IPSUR compilation.}
\source{IPSUR.RData was generated by the IPSUR build process.}
\keyword{datasets}
