\name{IPSUR-package}
\alias{IPSUR-package}
\alias{IPSUR}
\docType{package}
\title{
Introduction to Probability and Statistics Using R
}
\description{
An introductory probability and statistics textbook, alongside other supplementary materials.  The book is released under the GNU Free Documentation License.
}
\details{
\tabular{ll}{
Package: \tab IPSUR\cr
Type: \tab Package\cr
Version: \tab 4.0.1206\cr
Date: \tab 2022-12-06\cr
Depends: \tab R (>= 2.10)\cr
Imports: \tab actuar, aplpack, binom, boot, coin, distrEx, e1071, HH (>= 2.1-32), Hmisc, lmtest, mvtnorm, prob, qcc, RcmdrPlugin.IPSUR (>= 0.1-6), reshape, scatterplot3d, TeachingDemos (>= 2.5), vcd\cr
Suggests: \tab diagram, emdbook, ggplot2, lattice, RcmdrMisc\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
G. Jay Kerns
Maintainer: G. Jay Kerns <gkerns@ysu.edu>
}
\keyword{package}
