\name{IPSUR.Utilities}
\alias{IPSUR.Utilities}
\alias{read}

\title{IPSUR Utility Functions}
\description{
  These functions support using the IPSUR package.
}

\usage{read(x)}

\arguments{
  \item{x}{the name of a PDF in the doc directory, quoted or not.}
}

\details{
  These are convenience functions designed to make the use of IPSUR easier for novices.
}

\author{G. Jay Kerns \email{gkerns@ysu.edu}}

\keyword{misc}
