\name{PP}
\alias{PP}
\docType{data}
\title{Planned Parenthood data}
\description{
Data contains the services offered by the Planned Parenthood organization from the years 2006 to 2013.
}
\usage{data(PP)}
\format{
A multiple time series with on the following 6 variables.
  \describe{
  \item{Contraception}{contraceptive services}
  \item{STD.Test.Treat}{STD testing and treatment}
  \item{Cancer.Screen}{cancer screening services}
  \item{Preg.Prenat.Serv}{pregnancy and prenatal services}
  \item{Abortion.Serv}{abortion services}
  \item{Other.Serv}{other services}
  }
}
\source{Planned Parenthood annual reports, 2006-13.}
\keyword{datasets}
