\name{RcmdrIPSURPlugin-internal}
\title{Internal Rcmdr IPSUR Plugin objects}
\alias{barPlotSumTable}
\alias{barGraph.ipsur }
\alias{betaDistributionPlot.ipsur}
\alias{betaDistributionSamples.ipsur}
\alias{betaProbabilities.ipsur}
\alias{betaQuantiles.ipsur }
\alias{betaSimulate.ipsur }
\alias{binomialDistributionPlot.ipsur}
\alias{binomialDistributionSamples.ipsur}
\alias{binomialMass.ipsur}
\alias{binomialProbabilities.ipsur}
\alias{binomialQuantiles.ipsur }
\alias{binomialSimulate.ipsur }
\alias{birthdayProbabilities.ipsur}
\alias{boxPlot.ipsur}
\alias{cauchyDistributionPlot.ipsur}
\alias{CauchyDistributionSamples.ipsur}
\alias{cauchyProbabilities.ipsur}
\alias{cauchyQuantiles.ipsur }
\alias{cauchySimulate.ipsur }
\alias{chisquareDistributionPlot.ipsur}
\alias{chisquareDistributionSamples.ipsur}
\alias{chisqProbabilities.ipsur}
\alias{chisqQuantiles.ipsur }
\alias{chisqSimulate.ipsur }
\alias{disunifDistributionSamples.ipsur}
\alias{disunifSimulate.ipsur }
\alias{enterTable.ipsur}
\alias{enterTableMultiPropTest}
\alias{enterTableSinglePropTest}
\alias{expDistributionPlot.ipsur}
\alias{exponentialDistributionSamples.ipsur}
\alias{expProbabilities.ipsur}
\alias{expQuantiles.ipsur }
\alias{expSimulate.ipsur }
\alias{frequencyDistribution.ipsur}
\alias{FDistributionPlot.ipsur}
\alias{FDistributionSamples.ipsur}
\alias{FProbabilities.ipsur}
\alias{FQuantiles.ipsur }
\alias{fSimulate.ipsur }
\alias{gammaDistributionPlot.ipsur}
\alias{gammaDistributionSamples.ipsur}
\alias{gammaProbabilities.ipsur}
\alias{gammaQuantiles.ipsur }
\alias{gammaSimulate.ipsur }
\alias{geomDistributionPlot.ipsur}
\alias{geomDistributionSamples.ipsur}
\alias{geomMass.ipsur}
\alias{geomProbabilities.ipsur}
\alias{geomQuantiles.ipsur }
\alias{geomSimulate.ipsur }
\alias{hyperDistributionPlot.ipsur}
\alias{hyperDistributionSamples.ipsur}
\alias{hyperMass.ipsur}
\alias{hyperProbabilities.ipsur}
\alias{hyperQuantiles.ipsur }
\alias{hyperSimulate.ipsur }
\alias{logisticDistributionPlot.ipsur}
\alias{logisticDistributionSamples.ipsur}
\alias{logisticProbabilities.ipsur}
\alias{logisticQuantiles.ipsur }
\alias{logisSimulate.ipsur }
\alias{lognormalDistributionPlot.ipsur}
\alias{lognormalDistributionSamples.ipsur}
\alias{lognormalProbabilities.ipsur}
\alias{lognormalQuantiles.ipsur }
\alias{lnormalSimulate.ipsur }
\alias{multiSampleProportionsTest}
\alias{negbinomialDistributionPlot.ipsur}
\alias{negbinomialDistributionSamples.ipsur}
\alias{negbinomialMass.ipsur}
\alias{negbinomialProbabilities.ipsur}
\alias{negbinomialQuantiles.ipsur }
\alias{nbinomSimulate.ipsur }
\alias{normalDistributionPlot.ipsur}
\alias{normalDistributionSamples.ipsur}
\alias{normalProbabilities.ipsur}
\alias{normalQuantiles.ipsur }
\alias{normalSimulate.ipsur }
\alias{numericalSummaries.ipsur}
\alias{paretoChart}
\alias{PoissonDistributionPlot.ipsur}
\alias{PoissonDistributionSamples.ipsur}
\alias{PoissonMass.ipsur}
\alias{poissonProbabilities.ipsur}
\alias{poissonQuantiles.ipsur }
\alias{poisSimulate.ipsur }
\alias{powerTtest}
\alias{powerProptest}
\alias{powerAnovatest}
\alias{RcmdrEnv}
\alias{stripChart}
\alias{tDistributionPlot.ipsur}
\alias{tDistributionSamples.ipsur}
\alias{tProbabilities.ipsur}
\alias{tQuantiles.ipsur }
\alias{tSimulate.ipsur }
\alias{twoWayTable.ipsur}
\alias{unifDistributionPlot.ipsur}
\alias{uniformDistributionSamples.ipsur}
\alias{uniformProbabilities.ipsur}
\alias{uniformQuantiles.ipsur }
\alias{unifSimulate.ipsur }
\alias{UpdatebetasimNumber}
\alias{UpdatebinomsimNumber}
\alias{UpdatecauchysimNumber}
\alias{UpdatechisqsimNumber}
\alias{UpdatedisunifsimNumber}
\alias{UpdateexpsimNumber}
\alias{UpdatefsimNumber}
\alias{UpdategammasimNumber}
\alias{UpdategeomsimNumber}
\alias{UpdatehypersimNumber}
\alias{UpdatelnormsimNumber}
\alias{UpdatelogissimNumber}
\alias{UpdatenbinomsimNumber}
\alias{UpdatenormsimNumber}
\alias{UpdatepoissimNumber}
\alias{UpdatetsimNumber}
\alias{UpdateunifsimNumber}
\alias{UpdateweibullsimNumber}
\alias{weibullDistributionPlot.ipsur}
\alias{WeibullDistributionSamples.ipsur}
\alias{weibullProbabilities.ipsur}
\alias{weibullQuantiles.ipsur }
\alias{weibullSimulate.ipsur }



\description{Internal Rcmdr IPSUR Plugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}
