\name{RcmdrTestDrive}
\alias{RcmdrTestDrive}
\docType{data}
\title{ Test-driving the R Commander }
\description{
  These are simulated data specifically designed to allow the inexperienced user to browse the
  capabilities of the R Commander.
}
\usage{data(RcmdrTestDrive)}
\format{
  A data frame with 168 observations on the following 9 variables:


  \describe{
    \item{Order}{sequential order}
    \item{Smoking}{smoking status}
    \item{Gender}{gender of victim}
    \item{Race}{race of victim}
    \item{Before}{life expectancy before exposure}
    \item{After}{life expectancy after exposure}
    \item{Salary}{salary at retirement}
    \item{Reduction}{potential salary reduction}
    \item{Parking}{number of unpaid parking tickets}
  }
}
\details{
  The R Commander has extensive functionality, but many options are unavailable
  unless the correct types of data are loaded in the Active Data Set.  This
  data set was randomly generated so that, when loaded, essentially all R Commander
  options would be available for the student to investigate.  These data are
  entirely fictional.  For an amusing contributed story tying these variables together,
  please visit \url{http://www.cc.ysu.edu/~gjkerns/IPSUR/package}.
}

\keyword{datasets}
\source{
  These data were randomly generated using the IPSUR probability menu for the
  R Commander.

}


