\name{BloodPressure}
\alias{BloodPressure}
\docType{data}
\title{Blood Pressure and Heart Rate Readings}
\description{
  These data were collected during from 2004 through 2006
  by Taoying Bian.
}
\usage{data(BloodPressure)}
\format{
  A data frame with 202 observations on the following 7 variables:


  \describe{
    \item{year}{year. From 2004 through 2006}
    \item{month}{month of the year. January = 1.}
    \item{day}{the day of the month.}
    \item{hour}{the 24-clock hour.}
    \item{systolic}{systolic blood pressure reading (in mm Hg).}
    \item{diastolic}{diastolic blood pressure reading (in mm Hg).}
    \item{heart.rate}{heart rate reading, in beats per minute.}
  }
}
\details{
  From 2004 through 2006, Mrs. Taoying Bian regularly collected data concerning
  her blood pressure and heart rate.
}

\keyword{datasets}
\source{
  These data were collected by Taoying Bian from 2004 through 2006.
}


