\name{IPSUR-package}
\alias{IPSUR-package}
\alias{IPSUR}
\docType{package}
\title{
Introduction to Probability and Statistics Using R
}
\description{
This package contains the Emacs Org Mode source code used to generate IPSUR, an introductory probability and statistics textbook, alongside other supplementary materials such as the parsed R code for the book and searching functions for the HTML version.  The book is released under the GNU Free Documentation License.
}
\details{
\tabular{ll}{
Package: \tab IPSUR\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2011-09-16\cr
Depends: \tab R (>= 2.10)\cr
Suggests: \tab actuar, aplpack, boot, coin, combinat, diagram, distrEx, e1071, ggplot2, HH (>= 2.1-32), Hmisc, lattice, lmtest, mvtnorm, prob, qcc, RcmdrPlugin.IPSUR (>= 0.1-6), reshape, scatterplot3d, stats4, TeachingDemos (>= 2.5)\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
Repository: \tab R-Forge\cr
Repository/R-Forge/Project: \tab ipsur\cr
Repository/R-Forge/Revision: \tab 199\cr
Date/Publication: \tab 2011-03-24 18:05:34\cr
Packaged: \tab 2011-04-28 10:36:31 UTC; rforge\cr
}
}
\author{
G. Jay Kerns

Maintainer: G. Jay Kerns <gkerns@ysu.edu>
}
\keyword{ package }
