\name{IPSUR-package}
\Rdversion{1.1}
\alias{IPSUR-package}
\alias{IPSUR}
\docType{package}
\title{
Introduction to Probability and Statistics Using R
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab IPSUR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-09-03\cr
License: \tab file LICENSE\cr
LazyLoad: \tab yes\cr
}
An overview of how to use the package, including the most important stuff
}
\author{
G. Jay Kerns

Maintainer: G. Jay Kerns <gkerns@ysu.edu>

}
\references{
Literature or other references for background information
}
\keyword{ package }
