\name{IPSUR-package}
\Rdversion{1.1}
\alias{IPSUR-package}
\alias{IPSUR}
\docType{package}
\title{
Introduction to Probability and Statistics Using R
}
\description{
This is a self-contained textbook on probability and statistics for students with differential and integral calculus.  It is released under the GNU Free Documentation License.
}
\details{
\tabular{ll}{
Package: \tab IPSUR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-12-08\cr
License: \tab file LICENSE\cr
LazyLoad: \tab yes\cr
}
This package is a textbook written at the introductory level to introduce students to probability and statistics using R.  After loading the package with library(R), the book may be read by issuing the command read(IPSUR) at the command line. 
}
\author{
G. Jay Kerns

Maintainer: G. Jay Kerns <gkerns@ysu.edu>

}
\references{
Literature or other references for background information
}
\keyword{package}
