\name{IPSUR-package}
\Rdversion{1.1}
\alias{IPSUR-package}
\alias{IPSUR}
\docType{package}
\title{
Introduction to Probability and Statistics Using R
}
\description{
This package contains the Sweave source code used to generate IPSUR, an introductory probability and statistics textbook, alongside other supplementary materials such as the parsed R code for the book and data for the examples and exercises.  The book is released under the GNU Free Documentation License.
}
\details{
\tabular{ll}{
Package: \tab IPSUR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-12-08\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
This package is a textbook written at the introductory level to introduce students to probability and statistics using R.  After loading the package with library(IPSUR), the book may be read by issuing the command read(IPSUR) at the command line. 
}
\author{
G. Jay Kerns

Maintainer: G. Jay Kerns <gkerns@ysu.edu>

}

\keyword{package}
