\name{IPSUR-package}
\alias{IPSUR-package}
\alias{IPSUR}
\alias{ipsur}
\docType{package}
\title{
Introduction to Probability and Statistics Using R
}
\description{
  This package accompanies G. Andy Chang and G. Jay Kerns, Introduction to Probability and Statistics Using R (in progress).
  The package contributes functions unique to the book as well as specific configuration and selected functionality to the R Commander by John Fox.

}
\details{
\tabular{ll}{
Package: \tab IPSUR\cr
Version: \tab 0.1-3\cr
Date: \tab 2006-01-15\cr
Depends: \tab R (>= 2.1.0), car (>= 1.0-15), grDevices, tcltk, utils\cr
Suggests: \tab abind, distr, distrEx, effects (>= 1.0-7), foreign, grid, lattice, lmtest, MASS, mgcv, multcomp (>= 0.991-2), nlme, nnet, qcc, relimp, rgl, RODBC\cr
LazyLoad: \tab no\cr
License: \tab GPL version 2 or newer\cr
URL: \tab http://www.r-project.org,
http://www.cc.ysu.edu/~gjkerns/IPSUR/package\cr
}

}
\author{
G. Jay Kerns <gkerns@ysu.edu> with contributions by Theophilius Boye, adapted from the work of John Fox et al.

Maintainer: G. Jay Kerns <gkerns@ysu.edu>
}
\keyword{ package }
